requirements_ori.txt
appium-python-client==0.24
asn1crypto==0.22.0
bcrypt==3.1.3
blessings==1.6
cffi==1.10.0
colorama==0.3.9
coloredlogs==6.1
cryptography==1.9
cx-Oracle==5.1.2
decorator==3.3.3
docutils==0.9
enum34==1.1.6
humanfriendly==4.4.1
idna==2.5
ipaddress==1.0.18
Jinja2==2.9.6
MarkupSafe==1.0
monotonic==1.3
nose==1.3.7
nose-html-reporting==0.2.3
nosehtmloutput-2==1.0
olefile==0.44
paho-mqtt==1.3.0
paramiko==2.2.1
Pillow==4.2.1
ptyprocess==0.5.2
pyasn1==0.2.3
pycparser==2.17
pycrypto==2.6
PyNaCl==1.1.2
pypiwin32==219
PyQt4-windows-whl==4.11.3.2
pyqtconfig==0.8.6
pyserial==3.4
pywin32==221
requests==2.14.2
robotframework==3.0.2
robotframework-appiumlibrary==1.4.3
robotframework-difflibrary==0.1.0
robotframework-requests==0.4.7
robotframework-ride==1.5.2.1
robotframework-selenium2library==1.7.2
robotframework-sshlibrary==2.1.3
selenium==2.46.0
six==1.10.0
winpexpect==1.5
xlwt==1.2.0
xml==0.2.13
